Venda
Venda quase completo e produto arrumado

+++ src/main/java/mensal/service/VendaService.java
@@ -1,6 +1,7 @@
 package mensal.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,31 +12,41 @@ import mensal.repository.VendaRepository;
 @Service
 public class VendaService {
 	
-		@Autowired
-		private VendaRepository repository;
-		
-		public String save(Venda obj) {
-			this.repository.save(obj);
-			return "Venda salva com sucesso.";
-		}
-		
-		public List<Venda> listAll(){
-			return this.repository.findAll();
-		}
-		
-		public String update(long id, Venda obj) {
+	@Autowired
+	private VendaRepository repository;
+	
+	public String save(Venda obj) {
+		this.repository.save(obj);
+		return "Venda salva com sucesso.";
+	}
+	
+	public List<Venda> listAll(){
+		return this.repository.findAll();
+	}
+	
+	public String update(long id, Venda obj) {
+		Optional<Venda> optionalVenda = this.repository.findById(id);
+		if (optionalVenda.isPresent()) {
 			obj.setId(id);
 			this.repository.save(obj);
-			return "Venda não encontrada para alterar";
+			return "Venda atualizada com sucesso.";
+		} else {
+			return "Venda não encontrada para alterar.";
 		}
-		
-		public Venda findById(long id) {
-			Venda obj = this.repository.findById(id).get();
-			return obj;
-		}
-		
-		public String delete(long id) {
-			this.repository.deleteById(id);
-			return "Venda não encontrada para deletar";
-		}	
 	}
+	
+	public Venda findById(long id) {
+		Optional<Venda> optionalVenda = this.repository.findById(id);
+		return optionalVenda.orElse(null);
+	}
+	
+	public String delete(long id) {
+		Optional<Venda> optionalVenda = this.repository.findById(id);
+		if (optionalVenda.isPresent()) {
+			this.repository.deleteById(id);
+			return "Venda deletada com sucesso.";
+		} else {
+			return "Venda não encontrada para deletar.";
+		}
+	}	
+}
